+++ athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/GcsConstants.java
@@ -22,11 +22,12 @@ package com.amazonaws.athena.connectors.gcs;
 import com.amazonaws.athena.connector.lambda.QueryStatusChecker;
 import com.amazonaws.athena.connector.lambda.data.BlockSpiller;
 import com.amazonaws.athena.connector.lambda.records.ReadRecordsRequest;
+import com.amazonaws.athena.connectors.gcs.common.StorageSplit;
 
 public class GcsConstants
 {
     /**
-     * A deserialized JSON from an instance of {@link com.amazonaws.athena.connectors.gcs.storage.StorageSplit} to be added as a property
+     * A deserialized JSON from an instance of {@link StorageSplit} to be added as a property
      * of a Split. This Split will be passed to the {@link GcsRecordHandler#readWithConstraint(BlockSpiller, ReadRecordsRequest, QueryStatusChecker)} to
      * help know from which file it will read the records, along with record offset and total count of records to read
      */
